added some test scripts for appveyor, and moved some directories
