+++ .vs/restore.dg
@@ -1,3 +1,3 @@
-#:C:\Users\clark\Source\Repos\Mime-Detective\MimeDetective\MimeDetective.xproj
-#:C:\Users\clark\Source\Repos\Mime-Detective\MimeDetectiveTests\MimeDetectiveTests.xproj
-C:\Users\clark\Source\Repos\Mime-Detective\MimeDetectiveTests\MimeDetectiveTests.xproj|C:\Users\clark\Source\Repos\Mime-Detective\MimeDetective\MimeDetective.xproj
+#:C:\Users\clark\Source\Repos\Mime-Detective4\src\MimeDetective\MimeDetective.xproj
+#:C:\Users\clark\Source\Repos\Mime-Detective4\tests\MimeDetective.Tests.xproj
+C:\Users\clark\Source\Repos\Mime-Detective4\tests\MimeDetective.Tests.xproj|C:\Users\clark\Source\Repos\Mime-Detective4\src\MimeDetective\MimeDetective.xproj
