+++ MimeDetective.sln
@@ -3,9 +3,21 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "MimeDetective", "MimeDetective\MimeDetective.xproj", "{9E3066B2-B987-4E32-83D4-16A20E3BA219}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{098944FB-C1C9-48BE-AA37-CD3C5C336A84}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		appveyor.yml = appveyor.yml
+		Build.ps1 = Build.ps1
+		build.sh = build.sh
+		global.json = global.json
+		LICENSE = LICENSE
+		NuGet.config = NuGet.config
+		README.md = README.md
+	EndProjectSection
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "MimeDetectiveTests", "MimeDetectiveTests\MimeDetectiveTests.xproj", "{CB3A9BEA-C247-4FEB-8670-98E3034FC214}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "MimeDetective", "src\MimeDetective\MimeDetective.xproj", "{9E3066B2-B987-4E32-83D4-16A20E3BA219}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "MimeDetective.Tests", "tests\MimeDetective.Tests.xproj", "{CB3A9BEA-C247-4FEB-8670-98E3034FC214}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
