+++ appveyor.yml
@@ -1,0 +1,34 @@
+version: '{build}'
+skip_tags: true
+image: Visual Studio 2015
+configuration: Release
+install:
+  - ps: mkdir -Force ".\build\" | Out-Null
+  - ps: Invoke-WebRequest "https://raw.githubusercontent.com/dotnet/cli/rel/1.0.0-preview2/scripts/obtain/dotnet-install.ps1" -OutFile ".\build\installcli.ps1"
+  - ps: $env:DOTNET_INSTALL_DIR = "$pwd\.dotnetcli"
+  - ps: '& .\build\installcli.ps1 -InstallDir "$env:DOTNET_INSTALL_DIR" -NoPath -Version 1.0.0-preview2-003131'
+  - ps: $env:Path = "$env:DOTNET_INSTALL_DIR;$env:Path"
+build_script:
+- ps: ./Build.ps1
+test_script:
+  - nuget.exe install OpenCover -ExcludeVersion
+  - OpenCover\tools\OpenCover.Console.exe -register:user -filter:"+[MimeDetective]*" -target:"dotnet.exe" "-targetargs:test test\MimeDetective.Tests" -returntargetcode -hideskipped:All -output:coverage.xml
+  - "SET PATH=C:\\Python34;C:\\Python34\\Scripts;%PATH%"
+  - pip install codecov
+  - codecov -f "coverage.xml"
+artifacts:
+- path: artifacts/MimeDetective.*.nupkg
+deploy:
+- provider: NuGet
+  api_key:
+    secure: telB7SEpRsnG9gsi/cnGuxmNKPustCjeT+0FC4tuA9vQQcuQgx8yYWYAKZsl3YUR
+  skip_symbols: true
+  on:
+    branch: /^(master|dev)$/
+- provider: GitHub
+  auth_token:
+    secure: Fy6p3nueWv95xmQRo9Xbj5OgRcki2yZJ0IYJSAIvAi1D7pUI+0w01oy5Xzxe0HLc
+  artifact: /MimeDetective.*\.nupkg/
+  tag: v$(appveyor_build_version)
+  on:
+    branch: master
