+++ tests/project.json
@@ -1,0 +1,22 @@
+﻿{
+	"version": "1.0.0-*",
+
+	"testRunner": "xunit",
+	"dependencies": {
+		"xunit": "2.2.0-beta2-build3300",
+		"dotnet-test-xunit": "2.2.0-preview2-build1029",
+		"NETStandard.Library": "1.6.0",
+		"MimeDetective": "1.0.*"
+	},
+	"frameworks": {
+		"netcoreapp1.0": {
+			"dependencies": {
+				"Microsoft.NETCore.App": {
+					"type": "platform",
+					"version": "1.0.0"
+				}
+			}
+		}
+	}
+
+}
